// src/products/enums/product.enum.ts
export enum ProductEnum {
  ID_EXAMPLE = '2021a3a6-21c7-4010-8744-4c3c623fde61',
  ID_DESCRIPTION = 'Product ID',
  TITLE_EXAMPLE = 'Men’s Chill Crew Neck Sweatshirt',
  TITLE_DESCRIPTION = 'Product title',
  PRICE_EXAMPLE = '1500',
  PRICE_DESCRIPTION = 'Product price',
  DESCRIPTION_EXAMPLE = 'Introducing the Tesla Chill Collection',
  DESCRIPTION = 'Product description',
  SLUG_EXAMPLE = 'mens_chill_crew_neck_sweatshirt',
  SLUG_DESCRIPTION = 'Product slug',
  STOCK_EXAMPLE = '15',
  STOCK_DESCRIPTION = 'Product stock',
  SIZES_DESCRIPTION = 'Product sizes',
  GENDER_EXAMPLE = 'mem',
  GENDER_DESCRIPTION = 'Product gender',
  TAGS_DESCRIPTION = 'Product tags',
  IMAGES_DESCRIPTION = 'Product images url',
}